requirements.txt
base36==0.1.1
isodate==0.6.1
json_ref_dict==0.7.1
liftover==1.1.16
numpy==1.23.2
progress==1.6
pyhumps==3.8.0
pymongo==4.2.0
PyYAML==6.0
ruamel.yaml==1.0.0
